+++ index.d.ts
@@ -20,10 +20,12 @@ export interface AutoParseOptions {
   parseDates?: boolean;
   parseUrls?: boolean;
   parseFilePaths?: boolean;
+  onError?: (err: any, value: any, type?: any) => any;
   type?: any;
 }
 export type Parser = (value: any, type?: any, options?: AutoParseOptions) => any | undefined;
 export default function autoParse(value: any, typeOrOptions?: any | AutoParseOptions): any;
 export declare namespace autoParse {
   function use(fn: Parser): void;
+  function setErrorHandler(fn: ((err: any, value: any, type?: any) => any) | null): void;
 }